install_requirements.txt
transformers
torch
spacy 
pandas 
openai
bert-score
faiss-cpu
streamlit
openpyxl